docs(style): fix link visibility in dark mode (#2804)
fix: link visibility in dark mode

Hello, I noticed that links were using the same color as regular text making them indistinguishable in dark mode (default). This PR changes the color of links to a different color for better accessibility in dark more. This has been bugging me for a while, so I am creating this PR to fix this 😅.

Signed-off-by: Shivaditya Meduri <77324692+shivaditya-meduri@users.noreply.github.com>